requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.2
pandas==2.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
